+++ app/org/maproulette/models/dal/TaskDAL.scala
@@ -1532,7 +1532,7 @@ class TaskDAL @Inject()(override val db: Database,
           val pointParser = long("tasks.id") ~ str("tasks.name") ~ int("tasks.parent_id") ~ str("challenges.name") ~
             str("tasks.instruction") ~ str("location") ~ int("tasks.status") ~ get[Option[String]]("suggested_fix") ~
             get[Option[DateTime]]("tasks.mapped_on") ~ get[Option[Int]]("task_review.review_status") ~
-            get[Option[Int]]("task_review.review_requested_by") ~ get[Option[Int]]("task_review.reviewed_by") ~
+            get[Option[Long]]("task_review.review_requested_by") ~ get[Option[Long]]("task_review.reviewed_by") ~
             get[Option[DateTime]]("task_review.reviewed_at") ~ get[Option[DateTime]]("task_review.review_started_at") ~
             int("tasks.priority") ~ get[Option[Long]]("tasks.bundle_id") ~
             get[Option[Boolean]]("tasks.is_bundle_primary") map {
